refactor(app): Remove callbacks from FrontPage component
