+++ packages/openneuro-app/src/scripts/pages/front-page-comp.tsx
