+++ packages/openneuro-app/src/scripts/routes.tsx
@@ -6,7 +6,7 @@ import Dataset from './refactor_2021/dataset/draft-snapshot-routes'
 //import PreRefactorDatasetProps from './dataset/dataset-pre-refactor-container'
 
 import { FAQS } from '@openneuro/components/faqs'
-import FrontPageContainer from './refactor_2021/containers/front-page-container'
+import FrontPageContainer from './pages/front-page/front-page'
 import Admin from './pages/admin/admin'
 import SearchRoutes from './refactor_2021/search/search-routes'
 import APIKey from './refactor_2021/user/api'
