+++ packages/openneuro-components/src/front-page/index.ts
@@ -1,4 +1,3 @@
-import { FrontPage } from './FrontPage'
 import { AffiliateBlock } from './AffiliateBlock'
 import { Contributors } from './Contributors'
 import { GetUpdates } from './GetUpdates'
@@ -7,7 +6,6 @@ import { ActivityHeader } from './ActivityHeader'
 import { RecentData, TopViewed } from '../mock-content/activity-slider-content'
 
 export {
-  FrontPage,
   AffiliateBlock,
   Contributors,
   GetUpdates,
